Move broadcast channel to lib
