+++ src/lib/broadcast/index.ts
@@ -3,4 +3,9 @@ export default class BroadcastChannel {
   postMessage(_data: any) {}
   close() {}
   onmessage: (event: MessageEvent) => void = () => {}
+  addEventListener(_type: string, _listener: (event: MessageEvent) => void) {}
+  removeEventListener(
+    _type: string,
+    _listener: (event: MessageEvent) => void,
+  ) {}
 }
